A · P1 — 시장 채점이 도입 이후 한 건도 기록되지 않았다. $2 타입 모호.
## 증상 (운영 13:00 잡, 2026-09-05)

  WARNING [market] 채점 기록 실패 id=1: could not determine data type of parameter $2

  id 1~10 · 30~37 · 40~43 · 47 — **22건**이 같은 오류로 실패.

## 재현 (로컬 Postgres · 같은 스키마 · 같은 쿼리)

  종전 p_close=None   → AmbiguousParameterError: could not determine data type of parameter $2
  종전 p_close=0.512  → AmbiguousParameterError (동일)
  수정 p_close=None   → 통과  close=None  div=None  graded=True
  수정 p_close=0.512  → 통과  close=0.512 div=0.068 graded=True

**값과 무관하게 항상** 실패한다 — 프리페어 단계의 타입 추론 실패다.
그래서 2026-09-04 도입(e11cbaf) 이후 시장 채점이 **0건**이었고,
매일 13:00 에 `채점 기록 실패` 만 쌓였다.

## 원인

같은 `$2` 가 두 문맥에 쓰였다:
  SET p_market_close = $2       ← 컬럼 타입(NUMERIC)
  divergence = … our_p - $2     ← 산술
Postgres 가 둘을 통일하지 못한다.

## 수정 — 최소 침습

`$2::numeric` 캐스팅. **파라미터 개수를 바꾸지 않는다** — 호출부가
그대로면 되돌리기도 쉽고 회귀 표면이 좁다.

## 소급 — 코드가 필요 없다

`_PENDING` 이 `WHERE b.graded_at IS NULL` 로 고른다. 실패한 22건은
`graded_at` 이 여전히 NULL 이므로 **다음 `grade()` 에서 자동으로 다시
잡힌다.** 별도 백필 스크립트를 만들지 않는다 — 만들면 그 스크립트가
`grade()` 와 다른 규칙을 갖게 되는 순간 두 숫자가 갈린다.
→ 오늘 밤 13:00 잡(또는 다음 채점)에서 22건이 채워지는 것으로 확인한다.

[영향] ① 읽는 곳: market_baseline.grade 의 UPDATE 1곳. ② 상태: 실패해
       비어 있던 `p_market_close·divergence·market_hit·our_hit·graded_at`
       이 채워지기 시작한다. ③ 리그 분기 없음. ④ 실패는 종전대로
       `채점 기록 실패 id=…` 로 남는다 — 조용해지지 않았다.
       ⑤ 사본: 소급 규칙을 복제하지 않고 기존 대기 조건을 그대로 쓴다.

자기검증 (a) 실측만 — 재현 로그 첨부 (b) 4건 신설: 캐스팅 없는 `$2` 잔여
금지 · 파라미터 개수 불변 · 소급이 스크립트가 아님 · 사고 수치 보존
(c) 판정·발송 미변경. 1976 passed.

+++ app/engine/market_baseline.py
@@ -256,12 +256,22 @@ async def grade(pool, sport: str | None = None) -> dict:
             logger.debug("[market] our_hit 복사 실패 game=%s: %s", r["game_id"], exc)
             our_hit = None
         try:
+            # 🔴 [P1 2026-09-05] `$2` 를 **캐스팅한다.** 종전에는 같은 파라미터가
+            #    `SET p_market_close = $2`(컬럼 타입)와 `our_p - $2`(산술) 두
+            #    문맥에 쓰여 Postgres 가 타입을 통일하지 못했다:
+            #      asyncpg.exceptions.AmbiguousParameterError:
+            #      could not determine data type of parameter $2
+            #    값과 무관하게 **항상** 실패한다(재현 완료: p_close 가 None 이든
+            #    0.512 든 동일). 그래서 2026-09-04 이후 시장 채점이 **한 건도**
+            #    기록되지 않았고, 13:00 잡이 매번 `채점 기록 실패 id=…` 만 남겼다.
+            #    ⚠️ 파라미터 개수를 바꾸지 않는다 — 호출부 불변이 최소 침습이다.
             await pool.execute(
                 """UPDATE market_baseline_ledger
-                      SET p_market_close = $2, market_favored = COALESCE($3,
-                              market_favored),
-                          divergence = CASE WHEN our_p IS NOT NULL AND $2 IS NOT NULL
-                                            THEN our_p - $2 END,
+                      SET p_market_close = $2::numeric,
+                          market_favored = COALESCE($3, market_favored),
+                          divergence = CASE
+                              WHEN our_p IS NOT NULL AND $2::numeric IS NOT NULL
+                              THEN our_p - $2::numeric END,
                           market_hit = $4, our_hit = $5, void = $6,
                           graded_at = now()
                     WHERE id = $1""",
